# auth-service/auth-service-swagger-v1.yaml
openapi: 3.0.0
info:
  title: Auth Service API
  description: Authentication service for user login, registration, logout, and token management.
  version: 1.1.0

paths:
  /api/auth/register:
    post:
      summary: User registration
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: User successfully registered and tokens generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Registration failed.

  /api/auth/login:
    post:
      summary: User login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: User successfully logged in and tokens generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid credentials

  /api/auth/refresh:
    post:
      summary: Refresh JWT Token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
      responses:
        '200':
          description: Successfully refreshed the access token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid or expired refresh token.

  /api/auth/logout:
    post:
      summary: Logout from current device
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
      responses:
        '200':
          description: Successfully logged out.
          content:
            application/json:
              schema:
                type: string
                example: "Logged out"
        '400':
          description: Logout failed or invalid token.

  /api/auth/logout-all:
    post:
      summary: Logout from all devices
      operationId: logoutAll
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
      responses:
        '200':
          description: Successfully logged out from all devices.
        '400':
          description: Invalid refresh token.

components:
  schemas:
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    TokenRefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken

    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string