# auth-service/src/main/java/com/erayunal/dto/AuthResponse.java
package com.erayunal.dto;

import lombok.AllArgsConstructor;
import lombok.Data;

@Data
@AllArgsConstructor
public class AuthResponse {
    private String accessToken;
    private String refreshToken;
}